requirements.txt
Flask
requests
python-dotenv
flask-login
psycopg2-binary
Flask-SQLAlchemy==2.1